+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## 0.0.1
+## 1.1.0
 
 Collection of Widgets and helpful Methods that every flutter developer needs
-1. num extensions
+
+1. MScrollBehavior : custom scroll behavior class
+2. MTextFormField : custom text form field widget
+3. MCard : custom card widget
+4. MInkwell : custom inkwell widget
+5. MSlider : custom slider widget